requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
rich==13.7.0
prompt-toolkit==3.0.43
aiofiles==23.2.1
jinja2==3.1.2
pydantic==2.5.0
starlette==0.27.0
websockets==12.0
cryptography==41.0.8
pyjwt==2.8.0
requests==2.31.0
apscheduler==3.10.4
tiktoken==0.5.2
tiktoken-ext==0.0.3
numpy==1.24.4
psutil==5.9.6
Pillow==10.1.0